Feature/hot fix (#186)
* fixed

* update

+++ src/transaction/algo.ts
@@ -52,7 +52,7 @@ export const prepareAlgoSignedTransaction = async ( testnet: boolean, tx: AlgoTr
     const txn = {
         "from": tx.from,
         "to": tx.to,
-        "fee": BigInt(Number(tx.fee) * 1000000),
+        "fee": Number(tx.fee) * 1000000,
         "amount": BigInt(Number(tx.amount) * 1000000),
         "firstRound": params.firstRound,
         "lastRound": params.lastRound,
